requirements.txt
python-dotenv
google-generativeai
google
langchain
langchain_google_genai
langchain-community
streamlit
validators
-e .